requirements.txt
alembic==1.5.3
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
Pillow==8.1.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
qiniu==7.3.1
redis==3.5.3
requests==2.25.1
ronglian-sms-sdk==1.0
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
xmltodict==0.12.0
